OSSM-5584 Move integration tests into a separate package
